refactor: move activeExp into provider
