+++ src/types/searchItem.ts
@@ -1,0 +1,7 @@
+export type SearchItem = {
+  label: string;
+  content: string[];
+  value: string;
+  icon: JSX.Element;
+  shortcut: string;
+};
